Moved settings actions and reducers into parent action and reducers folders and updated settings files accordingly
